move getFrameGroups to utils so groupingComponent can use it
