+++ static/app/components/events/groupingInfo/groupingComponentStacktrace.tsx
@@ -4,12 +4,7 @@ import type {EventGroupComponent} from 'sentry/types/event';
 
 import GroupingComponent from './groupingComponent';
 import GroupingComponentFrames from './groupingComponentFrames';
-import {groupingComponentFilter} from './utils';
-
-type FrameGroup = {
-  data: EventGroupComponent[];
-  key: string;
-};
+import {getFrameGroups} from './utils';
 
 type Props = {
   component: EventGroupComponent;
@@ -24,33 +19,9 @@ function GroupingComponentStacktrace({
   onCollapsedChange,
   collapsed = false,
 }: Props) {
-  const getFrameGroups = () => {
-    const frameGroups: FrameGroup[] = [];
-
-    (component.values as EventGroupComponent[])
-      .filter(value => groupingComponentFilter(value, showNonContributing))
-      .forEach(value => {
-        const key = (value.values as EventGroupComponent[])
-          .filter(v => groupingComponentFilter(v, showNonContributing))
-          .map(v => v.id)
-          .sort((a, b) => a.localeCompare(b))
-          .join('');
-
-        const lastGroup = frameGroups[frameGroups.length - 1];
-
-        if (lastGroup?.key === key) {
-          lastGroup.data.push(value);
-        } else {
-          frameGroups.push({key, data: [value]});
-        }
-      });
-
-    return frameGroups;
-  };
-
   return (
     <Fragment>
-      {getFrameGroups().map((group, index) => (
+      {getFrameGroups(component, showNonContributing).map((group, index) => (
         <GroupingComponentFrames
           key={index}
           items={group.data.map((v, idx) => (
